[REFACTOR] refactor get issues API
[why]
For testability and maintainability

[what]
issue route
issue controller
issue service
issue model
assignee model
labelIssue model

[etc]
etc...
