+++ Backend/Models/assigneeModel.js
@@ -1,0 +1,8 @@
+const { db } = require('./dbPool');
+
+const getUserIdByIssueId = (issueId) => {
+  const query = 'SELECT userId FROM assignees WHERE issueId = ? ORDER BY userId ASC';
+  return db.execute(query, [issueId]);
+};
+
+module.exports = { getUserIdByIssueId };
