+++ Backend/Models/issueModel.js
@@ -1,0 +1,8 @@
+const { db } = require('./dbPool');
+
+const getById = (id) => {
+  const query = 'SELECT id, userId, title, milestoneId, isOpen, createdAt, openCloseAt FROM issues WHERE id = ?';
+  return db.execute(query, [id]);
+};
+
+module.exports = { getById };
